feat: allow customising context via callback on transport servers
This commit adds an callback function to the transport servers that allows
developers to inject context values into the server context.

This can be used to inject context values extracted from environment variables
(in stdio mode) or from headers (in sse mode), and access them in tools
using the provided context.