Add Callbacks Feature
MCPServer is upgraded with the capacity to notify before and/or
after requests run.  Developers can choose Before / After events,
and can submit callbacks that are notified during EVERY request
lifecycle, or else specific request methods.  The specific request
method callbacks receive typed message and/or result objects.

A special OnError callback is also provided.

To accomplish this with a minimal margin for error, a bit of a
refactoring was done:

- The `MCPServer.HandleMessage` method was moved from server.go
  into its own file.
- The individual request handlers' prototypes were changed.
  rather than returning `mcp.JSONRPCMessage`, they return a tuple
  of their specific result type and error.
- A code generation schema was introduced, to facilitate adding
  functionality to each clause in the switch statement within
  `HandleMessage()`.
- Now that HandleMessage() now receives a typed result object
  and possible error, it is able to pass those results to any
  configured callbacks.

I considered several options including a strategy pattern to make
HandleMessage() more DRY, but opted to stick with the minimal
reorganization of the code, and to keep the algorithm as flat
and obvious as possible.  I think the codegen mitigates the
repetitiveness of that section of code.