Export SendNotificationToAllClients
[First PR to project]

Export sendNotificationToAllClients so it can be called from the MCP
server to send unsolicited notifications to all clients (sessions).

I have a use-case for my MCP server where I need to send notifications
of state change to all sessions.  My MCP server is fronting a device
hub, and the device's on the hub would like to send notifications up
to the LLM.

There is probably a reason why this was not exported, so please forgive
this PR if that's the case.  I just could not figure out what the spec
says about this.

In any case, I added a test and tested with the Inspector.  The
Inspectory sees the unsolicited notifications.  Cool!  I also tried with
Cursor and Claude, but neither seem to recognize notifications, no
matter how hard I tried to convince them.