feat!: add context.Context to TokenStore methods
In distributed environments a TokenStore will be implemented as a
database, and often a context is required when running get/store
operations to ensure cancellation propagation, instrumentation,
etc works when running token store ops. This is a breaking change
but I think it makes sense (and it's pretty easy to fix
any breakages).