fix(tool): Return the `OutputSchema` from the tool definition
This commit returns the `OutputSchema` from the tool definition as per
the MCP spec: https://modelcontextprotocol.io/specification/2025-06-18/server/tools#output-schema

The behaviour of `WithOutputSchema()` also changed as `RawOutputSchema`
is not longer populated, but `OutputSchema` is from the `T` generic
type. The only way now to set `RawOutputSchema` is through the
`WithRawOutputSchema()` method.